+++ src/server/sv_game_java.c
@@ -202,8 +202,10 @@ SV_inPVS(const vec3_t p1, const vec3_t p2)
   cluster = CM_LeafCluster(leafnum);
   area2 = CM_LeafArea(leafnum);
 
-  if (mask && (!(mask[cluster >> 3] & (1 << (cluster & 7)))))
+  if (mask && (!(mask[cluster >> 3] & (BIT(cluster & 7)))))
+  {
     return qfalse;
+  }
 
   if(!CM_AreasConnected(area1, area2))
     return qfalse; //a door blocks sight
@@ -236,8 +238,10 @@ SV_inPVSIgnorePortals(const vec3_t p1, const vec3_t p2)
   cluster = CM_LeafCluster(leafnum);
   area2 = CM_LeafArea(leafnum);
 
-  if (mask && (!(mask[cluster >> 3] & (1 << (cluster & 7)))))
+  if (mask && (!(mask[cluster >> 3] & (BIT(cluster & 7)))))
+  {
     return qfalse;
+  }
 
   return qtrue;
 }